requirements.txt
streamlit
python-chess
torch
numpy
matplotlib
tqdm
pandas